Fetch All Data from Rick and Morty API - still in app.js not in service directory
