+++ src/components/App/App.css
@@ -2,4 +2,5 @@
   display: grid;
   gap: 40px;
   padding: 20px;
+  width: 350px;
 }
