+++ src/components/Card/Card.css
@@ -1,9 +1,12 @@
 .Card {
   background-color: lightcoral;
-  padding: 10px 30px 30px;
   box-shadow: 0px 5px 5px #aaa;
-  max-width: 90%;
   border-radius: 4px;
+  max-width: 350px;
+}
+
+.Card__header {
+  padding: 0 30px;
 }
 
 .Card__title {
@@ -23,13 +26,13 @@
 }
 
 .Card__detailsButton {
-  margin-bottom: 20px;
   width: 100%;
-  padding: 7px;
+  margin-bottom: 20px;
+  padding: 10px;
   border-radius: 4px;
   border: none;
   font-family: monospace;
-  font-size: 14px;
+  font-size: 16px;
   letter-spacing: 0.1em;
   text-transform: uppercase;
   color: #444;
@@ -37,5 +40,5 @@
 
 .Card__profilePic {
   width: 100%;
-  border-radius: 4px;
+  border-radius: 0 4px;
 }
