+++ src/components/Grid/Grid.css
@@ -1,0 +1,10 @@
+.Grid {
+  display: grid;
+  grid-template-rows: 48px auto 48px;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+}
